remove some modern attribute classes, closes #1751
Digging in the vault I discovered that already back in P5 version 1.9.0 the rendition attribute was a member of the att.global class. So, since it was not deleted explicitly before, I left it in this customisation. Second, I removed all 'modern' classes that were introduced after 1.9.0. Hope that's ok with the spirit of TEI bare?!