Addresses #1845.
And a lot besides! in attempting to rationalize the examples
containing witDetail, I found a number of related problems,
including a discussion of using @resp in a fashion now
deprecated in favor of @source. I also clarified the use of
the wit element vs. the @wit attribute and pared down some of
the rather intricate prose in TC.