type= on interpLike elements (#2053)
* NOT #2012:
Turns out we can’t fix #2012 the way we think it should be fixed because of Stylesheets bug 370. So here I have made an outright copy of the type= attribute we wanted to have in att.interpLike (and thus inherited by interp, interpGrp, span, and spanGrp) directly on interp, interpGrp, span, and spanGrp. I have included comments to explain what was going on, thus to make it easy to do the right thing when bug 370 is fixed. Each of this comments contains the string U+0023 370 for easier finding. (Although there may be a few false positives.)

* Use "replace" not "change":
Peter Stadler noticed that I had incorrectly used mode=change instead of mode=replace on the attribute definition of type= of spanGrp.

Co-authored-by: Syd Bauman <s.bauman@northeastern.edu>