tidy up detestable detest (#2268)
* Build: 'detest' improvements per #2267, step 1 (ostensibly uncontroversial):
Mostly this commit is the bulk of the work on the Makefile (and its subroutine, antruntest.xml) listed in #2267. The one thing *not* done is to separate the log file comparison into 3 separate small log file comparisons, as I fear that may be controversial.
NOTE: Most of the changes to detest.xml, and thus expected-results/detest.log, are the result of whitespace differences, I am afraid. (Probably due to starting with wrong or not updated branch when I made this branch, sorry.)
I am personally of the opinion that we should simply never have any TAB (U+09) characters at the begining of a line (i.e., before the 1st non-whitespace character on any given line) in files like detest.xml.

* Build: 'detest' improvements per #2267, step 2 (possibly controversial):
Separate the log file comparison into 3 separate small log file comparisons.
NOTE: It is likely that if & when this branch is merged into dev the actual Test/expected-results/*.log files will have to be changed

* Bug fix:
Oops! Forgot that Sebastion had removed some information lines from detest1 before comparison. Duplicated that idea in new process to avoid errors when comparing expected-results/ generated in one environment to output generated in another.

* Bug (& typo) fix:
Apparently I was using extended regexp syntax w/o telling grep I was doing so. Also discovered that 'fgrep' and 'egrep' are deprected, so used the preferred 'grep -F' and '-E' instead. Plus fixed a typo @martindholmes noticed.

* Fix wording of message